Refactored persistent volume controllers to new packaging structure
